SpaceFloat
Fix use of float in Chipmunk++
Use SpaceFloat for all physics values
Fix assigning float or int literals to SpaceFloat (currently double) value
Refactor int -> size_t for array index variables. This includes simplifying TextListMenuLayer to avoid wrap-around.

Fix mixing types in bound walls calculations
Fix ai::Flee::distance default init value

Add check for null equippedSpell in updateSpellControls()
Add check for null Player pointer in PlayerBatMode

+++ Classes/menu.h
@@ -63,7 +63,7 @@ private:
     vector<Label*> menuItemLabels;
     
     Node* cursor;
-    int selected = 0;
+    size_t selected = 0;
     
     void updateCursor()
     {
